List available SNIRH parameters

Returns a summary of available parameters in the conversion table, organized by sample type. This helps users understand what parameters can be converted to SNIRH format.

Usage

list_snirh_parameters(sample_type = "all", include_conversion_info = FALSE)

Arguments

sample_type

Character string specifying the sample type to filter by. Must be one of "water", "biota", "sediment", or "all". Default is "all".

include_conversion_info

Logical. If TRUE, includes conversion factors and unit information. Default is FALSE.

Value

A data.table with parameter information. Columns depend on include_conversion_info parameter.

Details

This function provides an overview of the parameter conversion capabilities of the package. It can help users:

Understand what parameters are supported

Check parameter naming conventions

Verify unit conversion factors

Plan data preparation activities

See also

Examples

# \donttest{
# List all water parameters
water_params <- list_snirh_parameters("water")
print(head(water_params))
#>    sample_type                 param_lab
#>         <char>                    <char>
#> 1:       biota       1,2,3,4,6,7,8-HpCDD
#> 2:       biota 1,2,3,4,6,7,8-HpCDD (TEQ)
#> 3:       biota       1,2,3,4,6,7,8-HpCDF
#> 4:       biota 1,2,3,4,6,7,8-HpCDF (TEQ)
#> 5:       biota       1,2,3,4,7,8,9-HpCDF
#> 6:       biota 1,2,3,4,7,8,9-HpCDF (TEQ)
#>                                              param_snirh
#>                                                   <char>
#> 1:               1,2,3,4,6,7,8-HpCDD (ng/kg peso húmido)
#> 2: 1,2,3,4,6,7,8-HpCDD (TEQ) (ng OMS-TEQ/kg peso húmido)
#> 3:               1,2,3,4,6,7,8-HpCDF (ng/kg peso húmido)
#> 4: 1,2,3,4,6,7,8-HpCDF (TEQ) (ng OMS-TEQ/kg peso húmido)
#> 5:               1,2,3,4,7,8,9-HpCDF (ng/kg peso húmido)
#> 6: 1,2,3,4,7,8,9-HpCDF (TEQ) (ng OMS-TEQ/kg peso húmido)
#>                symbol_snirh
#>                      <char>
#> 1:      1234678-HpCDD_BIOTA
#> 2: 1234678-HpCDD(TEQ)_BIOTA
#> 3:      1234678-HpCDF_BIOTA
#> 4: 1234678-HpCDF(TEQ)_BIOTA
#> 5:      1234789-HpCDF_BIOTA
#> 6: 1234789-HpCDF(TEQ)_BIOTA

# Get detailed conversion information
detailed_params <- list_snirh_parameters("water", include_conversion_info = TRUE)
print(head(detailed_params))
#>    sample_type                 param_lab      unit_lab
#>         <char>                    <char>        <char>
#> 1:       biota       1,2,3,4,6,7,8-HpCDD         ng/kg
#> 2:       biota 1,2,3,4,6,7,8-HpCDD (TEQ) ng OMS-TEQ/kg
#> 3:       biota       1,2,3,4,6,7,8-HpCDF         ng/kg
#> 4:       biota 1,2,3,4,6,7,8-HpCDF (TEQ) ng OMS-TEQ/kg
#> 5:       biota       1,2,3,4,7,8,9-HpCDF         ng/kg
#> 6:       biota 1,2,3,4,7,8,9-HpCDF (TEQ) ng OMS-TEQ/kg
#>                                              param_snirh unit_snirh
#>                                                   <char>     <char>
#> 1:               1,2,3,4,6,7,8-HpCDD (ng/kg peso húmido)      ng/kg
#> 2: 1,2,3,4,6,7,8-HpCDD (TEQ) (ng OMS-TEQ/kg peso húmido)      ng/kg
#> 3:               1,2,3,4,6,7,8-HpCDF (ng/kg peso húmido)      ng/kg
#> 4: 1,2,3,4,6,7,8-HpCDF (TEQ) (ng OMS-TEQ/kg peso húmido)      ng/kg
#> 5:               1,2,3,4,7,8,9-HpCDF (ng/kg peso húmido)      ng/kg
#> 6: 1,2,3,4,7,8,9-HpCDF (TEQ) (ng OMS-TEQ/kg peso húmido)      ng/kg
#>                symbol_snirh factor
#>                      <char>  <num>
#> 1:      1234678-HpCDD_BIOTA      1
#> 2: 1234678-HpCDD(TEQ)_BIOTA      1
#> 3:      1234678-HpCDF_BIOTA      1
#> 4: 1234678-HpCDF(TEQ)_BIOTA      1
#> 5:      1234789-HpCDF_BIOTA      1
#> 6: 1234789-HpCDF(TEQ)_BIOTA      1

# Check all available sample types
all_params <- list_snirh_parameters("all")
unique_types <- unique(all_params$sample_type)
print(paste("Available sample types:", paste(unique_types, collapse = ", ")))
#> [1] "Available sample types: biota, sediment, water"
# }